createJournalDocument {PredDrugInducedLiverInjury}	R Documentation

Create the exposure and outcome cohorts

Description

Create the exposure and outcome cohorts

Usage

createJournalDocument(
  resultDirectory,
  analysisId = 1,
  cohortIds,
  cohortNames,
  includeValidation = T
)

Arguments

resultDirectory
The directory containing the results (outputFolder)

analysisId
An integer specifying the model Analysis_Id used to create the document

cohortIds
A vector of cohort ids

cohortNames
A vector of cohort names

includeValidation
Whether to look in the validation folder and add any validation results

Details

This function will create the exposure and outcome cohorts following the definitions included in this package.

[Package PredDrugInducedLiverInjury version 1.0.6 Index]