createMultiPlpReport {PredDrugInducedLiverInjury}	R Documentation

createMultiPlpReport

Description

Creates a word document report of the prediction

Usage

createMultiPlpReport(
  analysisLocation,
  protocolLocation = file.path(analysisLocation, "doc", "protocol.docx"),
  includeModels = F
)

Arguments

analysisLocation
The location of the multiple patient-level prediction study

protocolLocation
The location of the auto generated patient-level prediction protocol

includeModels
Whether to include the models into the results document

Details

The function creates a word document containing the analysis details, data summary and prediction model results.

Value

A work document containing the results of the study is saved into the doc directory in the analysisLocation

[Package PredDrugInducedLiverInjury version 1.0.6 Index]