execute {PredDrugInducedLiverInjury}	R Documentation

Execute the Study

Description

Execute the Study

Usage

execute(
  databaseDetails,
  outputFolder,
  createProtocol = F,
  createCohorts = F,
  runDiagnostic = F,
  viewDiagnostic = F,
  runAnalyses = F,
  createValidationPackage = F,
  useHydra = F,
  skeletonVersion = "v0.0.1",
  analysesToValidate = NULL,
  packageResults = F,
  minCellCount = 5,
  viewShiny = F,
  onlyFetchData = F,
  sampleSize = NULL,
  logSettings
)

Arguments

databaseDetails
Database details created using PatientLevelPrediction::createDatabaseDetails()

outputFolder
Name of local folder to place results; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

createProtocol
Creates a protocol based on the analyses specification

createCohorts
Create the cohortTable table with the target population and outcome cohorts?

runDiagnostic
Runs a diagnostic of the T, O and tar settings for the cdmDatabaseSchema - can be used to check whether to change settings or whether the prediction may not do well.

viewDiagnostic
Opens a shiny app with the diagnostic results (run after runDiagnostic completes)

runAnalyses
Run the model development

createValidationPackage
Create a package for sharing the models

useHydra
Whether to use Hydra to create the validation package (requires hydra to be installed) or download the master version of the skeleton (requires internet access)

skeletonVersion
The version of the validation skeleton to use

analysesToValidate
A vector of analysis ids (e.g., c(1,3,10)) specifying which analysese to export into validation package. Default is NULL and all are exported.

packageResults
Should results be packaged for later sharing?

minCellCount
The minimum number of subjects contributing to a count before it can be included in packaged results.

viewShiny
View a shiny app with the results

onlyFetchData
Only fetch data for the analyses without fitting models. Setting this flag will overwrite your input provided to the runAnalyses and createCohorts parameters.

sampleSize
The number of patients in the target cohort to sample (if NULL uses all patients)

logSettings
The log setting PatientLevelPrediction::createLogSettings()

Details

This function executes the PredDrugInducedLiverInjury Study.

Examples

## Not run: 
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
                                             
 databaseDetails <- PatientLevelPrediction::createDatabaseDetails(
 connectionDetails = connectionDetails,
 cdmDatabaseSchema = cdmDatabaseSchema,
 cdmDatabaseName = cdmDatabaseName,
 tempEmulationSchema = tempEmulationSchema,
 cohortDatabaseSchema = cohortDatabaseSchema,
 cohortTable = cohortTable,
 outcomeDatabaseSchema = cohortDatabaseSchema,
 outcomeTable = cohortTable,
 cdmVersion = cdmVersion
 )  
 
 logSettings <- PatientLevelPrediction::createLogSettings(
 verbosity = "INFO",
 timeStamp = T,
 logName = 'skeletonPlp'
 )                                          

execute(databaseDetails = databaseDetails,
        outputFolder = "c:/temp/study_results", 
        createProtocol = T,
        createCohorts = T,
        runDiagnostic = F,
        viewDiagnostic = F,
        runAnalyses = T,
        createValidationPackage = T,
        useHydra = F,
        skeletonVersion = 'v1.0.1',
        packageResults = F,
        minCellCount = 5,
        viewShiny = F,
        sampleSize = 10000,
        logSettings = logSettings
        )

## End(Not run)

[Package PredDrugInducedLiverInjury version 1.0.6 Index]