packageResults {PredDrugInducedLiverInjury}	R Documentation

Package the results for sharing with OHDSI researchers

Description

Package the results for sharing with OHDSI researchers

Usage

packageResults(outputFolder, minCellCount = 5)

Arguments

outputFolder
Name of folder containing the study analysis results (/)

minCellCount
The minimum number of subjects contributing to a count before it can be included in the results.

Details

This function packages the results.

[Package PredDrugInducedLiverInjury version 1.0.6 Index]