populateShinyApp {PredDrugInducedLiverInjury}	R Documentation

Adds the results of the study into a shiny app

Description

Adds the results of the study into a shiny app

Usage

populateShinyApp(
  outputDirectory = "./ShinyApp",
  shinyDirectory,
  resultDirectory,
  minCellCount = 10,
  databaseName = "sharable name of development data"
)

Arguments

outputDirectory
Name of folder containing the shiny app

shinyDirectory
Location of a template shiny app (with no data)

resultDirectory
The location of the results from the study

minCellCount
The minimum number of subjects contributing to a count before it can be included in the results.

databaseName
A friendly name for the database

Details

This function adds the results of the study into a shiny app ready for sharing

[Package PredDrugInducedLiverInjury version 1.0.6 Index]