submitResults {PredDrugInducedLiverInjury}	R Documentation

Submit the study results to the study coordinating center

Description

Submit the study results to the study coordinating center

Usage

submitResults(outputFolder, privateKeyFileName, userName, remoteFolder = NULL)

Arguments

outputFolder
Name of local folder where the results were generated; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

privateKeyFileName
A character string denoting the path to an RSA private key.

userName
A character string containing the user name.

remoteFolder
The remote folder to upload the file to (default is 'patientLevelPrediction').

Details

This will upload the file StudyResults.zip to the OHDSI SFTP server This requires an active internet connection.

Value

TRUE if the upload was successful.

[Package PredDrugInducedLiverInjury version 1.0.6 Index]